Convert train_densenet.py to distributed data parallel
Note that I've confirmed this runs but have not checked timing vs single-node
data parallel, or single-gpu. To do so, the simplest way is to run each for 2
hours with a single node without the --val-iters argument. This way we will do
validation once per epoch and we can compare raw epoch timings. After that, we
can compare to DDP with 2,4,8,16, and 32 nodes to see how well we scale.