Add study and report loading to mimic_cxr_jpg.py
This also adds much more information in the module output. If you call
`python mimic_cxr_jpg.py` you can explore the size and label prevalences
for various choices of datasets.

Note that official_split() and cv() can each now output dataloaders if
requested.

Study dataloaders can be requested with `return_studies=True`, and
examples are shown in the output of how to use this. A dataframe
restricted to each study is shown, and they are collated across a batch
by stacking.

Also note that reports can be loaded into the study metadata in which
case they are repeated in the ReportText column for each row (dicom
image) in the study. These reports are loaded directly from the zip file
`mimic-cxr-reports.zip` using the Python zipfile module. There is
currently a limitation with that module that prevents thread-safety,
even with read-only access. In the examples, I use `num_workers=1` in my
DataLoaders, though note that this is likely to slow down training. An
alternative method would load the ReportText column directly into the
dataframe, but for very large datasets this may be impractical due to
memory limitations. Still another method would allow us to load the
extracted .txt files after first unzipping the reports file. This may be
ideal but would require another argument pointing to the subdirectory
holding the reports (we'd likely not want to maintain all the .txt files
while processing/downsampling, etc.).