Add train_densenet_ssl.py for semi-supervised learning
Note that I currently use a triplet loss between pixels, rows, or global
averaged features. I plan to extend this to optionally using a
contrastive loss. In order to evaluate this for transfer learning, I
plan to pretrain a number of feature networks, then update cxr_net in
train_densenet.py to be able to load the feature networks. Finally, I'll
compare supervised learning with the following initializations:
- from scratch
- from imagenet
- from SSL (with either scratch or imagenet pretraining itself)